Now the subscribers of newletter are saved in database and shown in dashboard
so the subscribers will be saved in database and shown in dashboard. js controlls if the data the user wanna submit is valid or not. Also if a subscriber has already submitted and tries to redo it with same email, in database i made email unique so that user cant do that and error alert wish show up